small changes to metab.bayesian.R
This may not have worked because there were several assignments <- that had nothing on the right-hand side. Also, I don't think my code needed the "rjags" library, so I commented it out for now. Several lines are currently in place for testing purposes (section "Read in data"), and will be deleted later.