kalman working – needs differential eq
kalman filter is “working” – but there is a problem with fitting
positive respiration (wrong sign given convention). When forced
negative, the fit behaves poorly. Either the lack of the differential
equation is playing a big role here, or there is a problem in how
uncertainty was propagated. Current method works well for trout bog day
319, but not on day 137.