metab wrapper, moved kalman, various edits
Wrote an initial version of the wrapper function, metab(). Still need
to handle NA’s in each of the metab functions (currently metab() throws
away a day if it has any NA’s), and needs to standardize inputs across
methods (or be adaptable) – currently works with metab.bayesian and
metab.kalman. Made more changes to addNAs, and to byeShort. Added a
tb.data.RData file to ext data – contains the output of ryanData().