added metab.mle, metabData, and small corrections
1) I added metab.mle, which has similar functionality to metal.optim. I
the nll function from the KFnllDO function, and wrapped it similar to
metab. I did this to create an mle proc. error function that had
input/output similar to the kalman and bayesian functions. 2) Changed
getSchmidt to not throw an error if an NA was present. 3) Changed
exponent in scale.exp.wind (is this right?). 4) Added metabData, which
functionally replaces ryanData (and is actually useful to other
people). metabData also references rLakeAnalyzer to calculate mixed
layer depth. metabData still needs to adaptable to other methods of
calculating k.gas.