Cleaned up Kalman
did away with the version of the Kalman filter where the loop ran in R (now just in C), and put the KalmanLoopR, KFnllDO, the metab.Kalman, and the KFSmoothDO functions all in the metab.kalman.R file. Included detailed comments on how the method works in the R portion of the code. C code contains little commenting about methods/ references to Harvey.