A Major overhaul of how functions are called on timeseries and how complexity is handled.
Details: All timeseries data adhere to a semi-strict format of a data.frame object with a "datetime" colum.
The column name uniquely identifies that data and the names used are semi-strict for calculations. For example,
k.cole, when applied to a data.frame, looks for a "wnd" column and does the calculation on that. It returns
a data.frame with datetime and k600 columns.

I will send out an email shortly about this.