wrote documentation for metab.mle; reuse eqs
the equations in the file are really solid, and can be recycled for metab.bayesian, metab.kalman, and somewhat for others. Includes both Latex and Ascii eqn and deqn. Good details section, right units, etc. Less focus on the actual coding aspect of the documentation, and more on the statistics.